+++ dctap/src/dctap.rs
@@ -31,10 +31,12 @@ impl DCTap {
         self.shapes.push(shape.clone());
     }
 
-    pub fn from_path(path: &Path, _config: TapConfig) -> Result<DCTap, TapError> {
+    pub fn from_path<P: AsRef<Path>>(path: P, _config: TapConfig) -> Result<DCTap, TapError> {
         let mut dctap = DCTap::new();
         debug!("DCTap parsed: {:?}", dctap);
-        let mut tap_reader = TapReaderBuilder::new().flexible(true).from_path(path)?;
+        let mut tap_reader = TapReaderBuilder::new()
+            .flexible(true)
+            .from_path(path, TapConfig::default())?;
         for maybe_shape in tap_reader.shapes() {
             let shape = maybe_shape?;
             dctap.add_shape(&shape)
@@ -45,7 +47,9 @@ impl DCTap {
     pub fn from_reader<R: io::Read>(reader: R) -> Result<DCTap, TapError> {
         let mut dctap = DCTap::new();
         debug!("DCTap parsed: {:?}", dctap);
-        let mut tap_reader = TapReaderBuilder::new().flexible(true).from_reader(reader)?;
+        let mut tap_reader = TapReaderBuilder::new()
+            .flexible(true)
+            .from_reader(reader, TapConfig::default())?;
         for maybe_shape in tap_reader.shapes() {
             let shape = maybe_shape?;
             dctap.add_shape(&shape)