+++ rdfsx_cli/src/cli.rs
@@ -26,6 +26,7 @@ pub struct Cli {
 
 #[derive(Subcommand, Debug)]
 pub enum Command {
+    /// Information about ShEx shapemaps
     Shapemap {
         #[arg(short = 'm', long = "shapemap", value_name = "ShapeMap file name")]
         shapemap: PathBuf,
@@ -54,7 +55,8 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
-    Schema {
+    /// Information about ShEx schemas
+    ShEx {
         #[arg(short = 's', long = "schema", value_name = "Schema file name")]
         schema: PathBuf,
 
@@ -87,7 +89,8 @@ pub enum Command {
         )]
         output: Option<PathBuf>,
     },
-
+  
+    /// RDF Validation using ShEx schemas
     ValidateShex {
         #[arg(short = 's', long = "schema", value_name = "Schema file name")]
         schema: PathBuf,
@@ -150,6 +153,7 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
+    /// RDF Validation using Shacl shapes
     ValidateShacl {
         #[arg(short = 's', long = "shapes", value_name = "Shapes file name")]
         shapes: PathBuf,
@@ -181,6 +185,7 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
+    /// Information about RDF data
     Data {
         #[arg(short = 'd', long = "data", value_name = "RDF data path")]
         data: PathBuf,
@@ -201,6 +206,7 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
+    /// Information about RDF nodes which are part of RDF Graphs
     Node {
         #[arg(short = 'n', long = "node")]
         node: String,
@@ -241,6 +247,7 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
+    /// Information about SHACL shapes
     Shacl {
         #[arg(short = 's', long = "shapes", value_name = "Shapes file name")]
         shapes: PathBuf,
@@ -269,6 +276,7 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
+    /// Information and processing of DCTAP files
     #[command(name = "dctap")]
     DCTap {
         #[arg(short = 'd', long = "data", value_name = "DCTap file name")]
@@ -298,6 +306,7 @@ pub enum Command {
         output: Option<PathBuf>,
     },
 
+    /// Conversion between different Data modeling technologies
     #[command(name = "convert")]
     Convert {
         #[arg(short = 'm', long = "input-mode", value_name = "Input mode")]
@@ -571,6 +580,8 @@ pub enum OutputConvertFormat {
     Turtle,
     PlantUML,
     HTML,
+    SVG,
+    PNG,
 }
 
 impl Display for OutputConvertFormat {
@@ -584,6 +595,8 @@ impl Display for OutputConvertFormat {
             OutputConvertFormat::Turtle => write!(dest, "turtle"),
             OutputConvertFormat::PlantUML => write!(dest, "uml"),
             OutputConvertFormat::HTML => write!(dest, "html"),
+            OutputConvertFormat::PNG => write!(dest, "png"),
+            OutputConvertFormat::SVG => write!(dest, "svg"),
         }
     }
 }