+++ shapes_converter/src/shex_to_html/mod.rs
@@ -1,4 +1,4 @@
-//! ShEx to SPARQL
+//! ShEx to HTML
 //!
 //!
 mod cardinality;
@@ -10,6 +10,7 @@ mod node_id;
 pub mod shex2html;
 pub mod shex2html_config;
 pub mod shex2html_error;
+pub mod value;
 mod value_constraint;
 
 pub use cardinality::*;
@@ -20,4 +21,5 @@ pub use name::*;
 pub use node_id::*;
 pub use shex2html_config::*;
 pub use shex2html_error::*;
+pub use value::*;
 pub use value_constraint::*;