+++ shapes_converter/src/shex_to_uml/shex2uml_config.rs
@@ -1,12 +1,28 @@
-use std::{fs, io};
+use std::{
+    env::{self, VarError},
+    fs, io,
+    path::{Path, PathBuf},
+};
 
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-#[derive(Serialize, Deserialize, PartialEq, Debug, Default)]
-pub struct ShEx2UmlConfig {}
+pub const PLANTUML: &str = "PLANTUML";
+
+#[derive(Serialize, Deserialize, PartialEq, Debug)]
+pub struct ShEx2UmlConfig {
+    pub plantuml_path: Option<PathBuf>,
+}
 
 impl ShEx2UmlConfig {
+    pub fn new() -> ShEx2UmlConfig {
+        let plantuml_path = match env::var(PLANTUML) {
+            Ok(value) => Some(Path::new(value.as_str()).to_path_buf()),
+            Err(_) => None,
+        };
+        Self { plantuml_path }
+    }
+
     pub fn from_file(file_name: &str) -> Result<ShEx2UmlConfig, ShEx2UmlConfigError> {
         let config_str =
             fs::read_to_string(file_name).map_err(|e| ShEx2UmlConfigError::ReadingConfigError {
@@ -20,6 +36,11 @@ impl ShEx2UmlConfig {
             }
         })
     }
+
+    pub fn with_plantuml_path<P: AsRef<Path>>(mut self, path: P) -> Self {
+        self.plantuml_path = Some(path.as_ref().to_owned());
+        self
+    }
 }
 
 #[derive(Error, Debug)]
@@ -32,4 +53,13 @@ pub enum ShEx2UmlConfigError {
         path_name: String,
         error: serde_yaml::Error,
     },
+
+    #[error("Accessing environment variable {var_name}: {error}")]
+    EnvVarError { var_name: String, error: VarError },
+}
+
+impl Default for ShEx2UmlConfig {
+    fn default() -> Self {
+        Self::new()
+    }
 }